requirements.txt
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
google-api-core==2.10.2
google-auth==2.12.0
google-cloud-vision==3.1.4
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
idna==3.4
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.7
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyQt6==6.4.0
PyQt6-Qt6==6.4.0
PyQt6-sip==13.4.0
pytest==7.1.3
pytest-cov==4.0.0
requests==2.28.1
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
